Primarily an update to meta_runner.py
* Introduced parallelisation inside `meta_runner` (currently only for
bounded testing), which takes care of computing coverage and emitting
the logged data correctly
* Rewrote how logging is done (removed ugly hardcoding from before)
* Now pulling `gcovr` in the Dockerfile
* Slight update to `isl-meta-stats.h` helper file regarding print
formatting

+++ include/testing/isl-meta-stats.h
@@ -1,7 +1,7 @@
 #ifndef ISL_META_STATS_HPP
 #define ISL_META_STATS_HPP
 
-static unsigned int c_cnt;
+static size_t c_cnt;
 
 isl::stat
 getBSConstraints(isl::basic_set bs)
@@ -20,7 +20,7 @@ printStats(isl::set s)
 	fprintf(stderr, "SET EMPTY = %s\n", s.is_empty() ? "true" : "false");
     fprintf(stderr, "N BASIC SET = %d\n", s.n_basic_set());
     s.foreach_basic_set(getBSConstraints);
-    fprintf(stderr, "N CONSTRAINTS = %d\n", c_cnt);
+    fprintf(stderr, "N CONSTRAINTS = %zu\n", c_cnt);
 }
 
 #endif
